Implement a simple composer activation mechanism
We can now use `i` or `I` to enter the composer from a Kodachi "window."
The composer will hide itself (but not lose its text!) when leaving it.

Future work could support VTT modes (where the server controls echo)
when hitting `i` by remapping every key and sending it directly to the
server.