Add basic Trigger support (#2)
* Prepare lua API for declaring triggers

* Receive TriggerFired events

* Fix: not filtering by connection_id

* Notify daemon that we're "clearing" state

* Update MatcherSpec types to ensure rust casing

* Scaffold out rust handling of Matchers/TextProcessors

This commit adds the `bytes` crate to facilitate byte array
manipulation, for text matching and extraction purposes. Some triggers
(primarily, prompts) will want to "consume" matched strings so they
don't get printed to the output.

* Prepare logic and utilities for handling Triggers

* Implement core trigger processing logic

* Clean up accomplished TODO

* Introduce a macro to simplify ansi writing

* Scaffold matcher "compilation" step + basic match checking

* Implement Clear command

* Add basic match extraction

* Scaffold out trigger match notification

* Refactor match context storage to be serializable; return to client

* Refactor Ansi to abandon notion of deref'ing into ansi-stripped

We now have 3 classes: Ansi, AnsiMut, and AnsiStripped. AnsiMut is the
old, mutable Ansi from before this commit. Ansi is immutable, and uses
Bytes under the hood to be trivially clone()'able, and AnsiStripped
wraps the text with Ansi sequences removed.

The next step will be to encode data about the stripped Ansi sequences'
locations, so AnsiStripped can be used to reverse-map ranges to the
original Ansi string.

If possible, it'd be cool if strip_ansi() could create AnsiStripped
without any byte copying, using slices from Bytes...

* Use split() to make take_bytes() more efficient

* Add basic support for mapping ranges from AnsiStripped back to Ansi

* Clean up resolved TODO

* Use a byte literal for readability

* Pass Ansi to TextProcessor

It should be unambiguous that it might receive text with Ansi sequences
in it.

* Fix hang caused by incomplete processing

This also finally refactors to "emit" results via a caller-provided Vec,
which means we no longer have to pass in the CHannel to the
TextProcessor---much cleaner.

* Clean up and unify IDs, etc. Triggers work!!

We now have a new `Notification` wrapper type for Notifications that
requires a unified `connection_id` property. Any non-connection-bound
notifications can use a new `Notification::OutOfBand` type, if we ever
need that.

* Rename Request.connection -> connection_id

Match responses/notifications

* Refactor with explicit Request/Notification client event types

This fixes the crash when running the Clear notification

* Clean up debug code