Decouple TextProcessor from output + enable more complicated UI
This refactor introduces a ProcessorOutputReceiver trait that
TextProcessor uses to communicate its progress in a more declarative
way, without having to know the details of how the output will be
rendered. It should also empower more stateful output mechanisms, like
(potentially) maintaining a persistent prompt at the bottom of the
screen.