Refactor so shared state is individually lock'able
UI State is now shared between the Receiver and the Connection, and
since RPC events are handled asynchronously from network events, it's
important that we can lock different pieces of state individually.
Otherwise, while processing network input in the TextProcessor, we would
be unable to access UI state (since Mutex is not re-entrant, and the
shared_state field was already locked).