Refactor Transport to be async with custom Telnet (#9)
* Decouple Transport trait from telnet library

* WIP: Begin rewriting telnet from scratch for async + minimal copy

* Implement very basic telnet protocol parsing

This should be sufficient to just completely ignore all the telnet
stuff, for now, and gives us a nice async API that brings our idle CPU
usage down to zero without having to rely on sleep!

* Clean up (now-)unnecessary data copy

* Handle literal IACs within a Subnegotiate

* Properly support a TLS TelnetTransport

* Fix: treating an incomplete Subnegotiation as Data

* Upgrade tokio

* Fix disconnect handling

I would've expected read_buf to return an Err on EOF, but instead it
returns `Ok(0)` which *could also* mean the buffer is full. So we check
to make sure the buffer *isn't* full and, if not, treat that as EOF.

* Reuse existing idle RPC process for a new connection

Primarily this fixes the issue where we're currently just spawning a new
process every time and abandoning the old one as a zombie in the
background, but this is how the RPC process was meant to be used anyway.