Introduce a CompletionSource trait that returns an Iterator of &String
Using an Iterator instead of just a Vec (esp for the reference, instead
of cloning each item!) is a bit tricky to implement but I hope it will
be worth it. This will let us apply result filtering at a higher level,
and also avoid having to clone every potential result just to drop it at
the RPC layer.

Mostly I just was curious if it was possible, so this was a learning
exercise! This was like my 3rd or 4th attempt at doing this; the key
insight was not to try to capture the Iterator (since the lifetimes are
*even trickier* there) but to wrap a reference to the Transitions and
build the iterator using the IntoIterator trait from the wrapped type.