Add fancier auto-completion (#15)
* Begin work on a MarkovTrie to power smarter output-based completions

This is largely based on the same collection I created for iaido, but
has been optimized somewhat to reduce the excessive amount of `clone()`
I was doing. Learning is fun!

* Refactor to not convert to VecDeque anymore

We can still avoid unnecessary cloning when visiting an existing node by
just being a bit more verbose

* Fix find_node not stopping soon enough

* Clean up a TODO

Does not seem to work, since the slice needs to have a Sized type

* Fix: max_depth ignored

* Add default stop words

* Add a rough MarkovCompletionSource implementation

* Introduce a CompletionSource trait that returns an Iterator of &String

Using an Iterator instead of just a Vec (esp for the reference, instead
of cloning each item!) is a bit tricky to implement but I hope it will
be worth it. This will let us apply result filtering at a higher level,
and also avoid having to clone every potential result just to drop it at
the RPC layer.

Mostly I just was curious if it was possible, so this was a learning
exercise! This was like my 3rd or 4th attempt at doing this; the key
insight was not to try to capture the Iterator (since the lifetimes are
*even trickier* there) but to wrap a reference to the Transitions and
build the iterator using the IntoIterator trait from the wrapped type.

* Add the DuplexSelectorFactory trait and a WeightedRandom implementation

Associated types is super helpful here! It means we don't have to
specify the selector type in the generics for our
DuplexCompletionSource, just the Factory's type itself.

* Add a test for WeightedRandomSelectorFactory

* Add the WordIndexSelectorFactory, with tests

* Reimplement Completions using DuplexCompletionSource

Look ma, no Box!

* Return an impl iterator from Completions.suggest

* Clean up type reference for readability

* Integrate CompletionParams-based filtering

We do *very* light candidate filtering and allow clients to do whatever
extra filtering they prefer on top.

* Clean up some spurious logging

* Simplify some CompletionParams instantiations with the new factories

* Handle querying a non-existent path

* Update README