Add support for "simple" matchers and Alias replacements (#16)
* Add initial support for compiling a regex from a "simple" matcher

* Handle `${var}` syntax

* Add Formatter, sharing simple alias logic to build strings

* Support literal $ symbols

* Wire up Formatter as an option for RegisterAlias

I decided to use a single event instead of muddying the waters with
RegisterSimpleAlias. It's kind of "six of one half dozen of the other"
here: the old way was explicit but verbose and annoying to maintain;
this way is simpler for most clients, and less code to maintain in the
RPC process, but the protocol is slightly muddier.

We *could* instead do something like MatcherSpec where we actually have
a `type` property. That keeps some explicitness without adding too much
code duplication. For whatever reason, it feels a bit bleh to me right
now, however....