Refactor output receiver impl to work more consistently
The existing version is messy, confused, and frankly a bit broken. The
basic case of appending text to a partial line isn't handled correctly,
and the flow of events to the `ExternalUI` is similarly broken.

This refactor simplifies things with a fairly straightforward flow of
events, and simultaneously fixes the issues I was running into before.

Proper write-up to come, probably in the PR