Provide the "plain" text to external UIs
Pre-processing the ANSI like this is much faster in Rust, so is very
convenient for external clients---it takes search within a somewhat
large buffer (>1k lines) from being janky and occasionally causing input
confusion (we should figure this out, separately) to actually useable.

We may eventually want to also support breaking up the chars into chunks
so we don't have to do *any* ansi parsing in external apps---might
provide a noticeable speedup for saya when receiving a very ansi-rich
line (like the "splash" screen at Procedural Realms).

We probably will want to make this configurable, so I've marked the type
as Option; for now, we always send it.